requirements.txt
django
django-drf-filepond
uuid
django-mathfilters
pycrypto
djangorestframework
djoser
Pillow
django-imagekit
geoip2
redis 
celery